move related components and pages to same folder
